[interfaces.smooth-function] add SmoothFunctionInterface
Following our paper, we have one strong assumption on (localizable)
functions:

  Once localized to an entity, every function _should_ be polynomial
  w.r.t. the evaluation on that entity and _has_ to provide the order
  which is required to integrate it exactly.

The main reason for this is that we always end up doing quadratures
and we are thus implicitly restricted to locally polynomial
approximations of data functions. This design decision forces the
user to think about this early, when defining a function.

Bottom line: the only valid reason to evaluate a function in global
coordinates is because it is a smooth function (in the C^\infty
sense). Hence, the old GlobalFunctionInterface is replaced by this
SmoothFunctionInterface. All other functions may be double-valued on
intersections and thus need to be localizable.