[interfaces.smooth-function] add as_localizable()
This adds support to _convert_ a smooth function into a
localizable one, as in:

```
const auto& localizable_func = smooth_func.as_localizable<E>();
```

Here, `E` is the type of the entity, w.r.t. which the
`localizable_func` should be localizable. To avoid temporaries when
passing this the result to other functions and classes,
`as_localizable` returns a const reference, not a value. Therefore
we store all returned wrappers in a local static map of wrappers.